[MOVE] GameService 에서 팀과 관련된 메서드를 GameTeamService 로 이동 #47
