+++ src/main/java/com/sports/server/game/application/GameService.java
@@ -4,11 +4,8 @@ import com.sports.server.common.exception.ExceptionMessages;
 import com.sports.server.common.exception.NotFoundException;
 import com.sports.server.game.domain.Game;
 import com.sports.server.game.domain.GameRepository;
-import com.sports.server.game.domain.GameTeam;
-import com.sports.server.game.domain.GameTeamRepository;
 import com.sports.server.game.dto.response.GameDetailResponseDto;
 import com.sports.server.game.dto.response.GameResponseDto;
-import com.sports.server.game.dto.response.GameTeamCheerResponseDto;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -20,14 +17,13 @@ import org.springframework.transaction.annotation.Transactional;
 public class GameService {
 
     private final GameRepository gameRepository;
-    private final GameTeamRepository gameTeamRepository;
 
     public GameDetailResponseDto getOneGame(final Long gameId) {
         Game game = findGameWithId(gameId);
         return new GameDetailResponseDto(game);
     }
 
-    private Game findGameWithId(final Long gameId) {
+    public Game findGameWithId(final Long gameId) {
         return gameRepository.findById(gameId)
                 .orElseThrow(() -> new NotFoundException(ExceptionMessages.GAME_NOT_FOUND_EXCEPTION));
     }
@@ -37,12 +33,4 @@ public class GameService {
                 .toList();
     }
 
-
-    public List<GameTeamCheerResponseDto> getCheerCountOfGameTeams(final Long gameId) {
-        Game game = findGameWithId(gameId);
-        List<GameTeam> gameTeams = gameTeamRepository.findAllByGame(game);
-        return gameTeams.stream()
-                .map(GameTeamCheerResponseDto::new)
-                .toList();
-    }
 }
